Split and Sort specialfunctions
This is in preparation for bigger refactoring.
